Git commit message colorisation support
Scopes for Git commit message and Diff seems to be different. I added the same rules for the git commit message scope so they can be separated. Previous PR wasn't using the parent scope so it was working for both Diff and Git commit message (which is diff again but its under different scope). Again :) feel free to modify/refine where you feel like. I really like your theme and the one that is not working is this Git commit message syntax and I use it a lot. Thanks!